Avoid ambiguity error on GCC

+++ crnlib/crn_vector.cpp
@@ -20,8 +20,8 @@ bool elemental_vector::increase_capacity(uint min_new_capacity, bool grow_hint, 
     return true;
 
   ptr_bits_t new_capacity = min_new_capacity;
-  if ((grow_hint) && (!math::is_power_of_2(new_capacity)))
-    new_capacity = math::next_pow2(new_capacity);
+  if ((grow_hint) && (!math::is_power_of_2((uint64)new_capacity)))
+    new_capacity = math::next_pow2((uint64)new_capacity);
 
   CRNLIB_ASSERT(new_capacity && (new_capacity > m_capacity));
